+++ bootique-linkmove3/src/main/java/io/bootique/linkmove/v3/LinkMoveFactory.java
@@ -19,17 +19,13 @@
 
 package io.bootique.linkmove.v3;
 
-import com.nhl.link.move.connect.Connector;
 import com.nhl.link.move.resource.FolderResourceResolver;
 import com.nhl.link.move.resource.ResourceResolver;
 import com.nhl.link.move.runtime.LmRuntime;
 import com.nhl.link.move.runtime.LmRuntimeBuilder;
+import com.nhl.link.move.runtime.connect.IConnectorFactory;
 import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
-import io.bootique.di.Injector;
-import io.bootique.linkmove.v3.connector.IConnectorFactoryFactory;
-import io.bootique.linkmove.v3.connector.JdbcConnectorFactoryFactory;
-import io.bootique.linkmove.v3.connector.URIConnectorFactoryFactory;
 import io.bootique.linkmove.v3.resource.BQUrlResourceResolver;
 import io.bootique.resource.FolderResourceFactory;
 import org.apache.cayenne.configuration.server.ServerRuntime;
@@ -37,11 +33,6 @@ import org.apache.cayenne.configuration.server.ServerRuntime;
 import java.io.File;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
@@ -52,65 +43,34 @@ import java.util.Set;
 public class LinkMoveFactory {
 
     private FolderResourceFactory extractorsDir;
-    private List<IConnectorFactoryFactory<? extends Connector>> connectorFactories;
-
-    public LinkMoveFactory() {
-        this.extractorsDir = new FolderResourceFactory(".");
-        this.connectorFactories = new ArrayList<>();
-
-        // add default factories
-        connectorFactories.add(new JdbcConnectorFactoryFactory());
-        connectorFactories.add(new URIConnectorFactoryFactory());
-    }
 
     public LmRuntime createLinkMove(
-            Injector injector,
             ServerRuntime targetRuntime,
+            Set<IConnectorFactory<?>> connectorFactories,
             Set<LinkMoveBuilderCallback> builderCallbacks) {
 
         ResourceResolver resolver = createResolver();
-        LmRuntimeBuilder builder = new LmRuntimeBuilder()
+        LmRuntimeBuilder builder = LmRuntime.builder()
                 .withTargetRuntime(targetRuntime)
                 .extractorResolver(resolver);
 
-        validateConnectorFactories();
-
-        connectorFactories.forEach(factory -> addToBuilder(builder, factory, injector));
+        connectorFactories.forEach(builder::connectorFactory);
 
         builderCallbacks.forEach(c -> c.build(builder));
         return builder.build();
     }
 
-    private void validateConnectorFactories() {
-        // LM allows only a single connector factory per type. So e.g. "uri" and "bq.rest" factories will override each
-        //  other. Let's validate this condition to avoid confusion, when suddenly a preconfigured connector becomes
-        // unavailable
-
-        if (connectorFactories.size() > 1) {
-            Map<Class, IConnectorFactoryFactory> seen = new HashMap<>();
-            for (IConnectorFactoryFactory<?> factory : connectorFactories) {
-                IConnectorFactoryFactory<?> seenFactory = seen.put(factory.getConnectorType(), factory);
-                if (seenFactory != null && seenFactory != factory) {
-                    throw new IllegalArgumentException("Two factories present for connector type '"
-                            + factory.getConnectorType().getName()
-                            + ": (1) "
-                            + seenFactory.getClass().getName()
-                            + ", (2) "
-                            + factory.getClass().getName());
-                }
-            }
-        }
-    }
-
     protected ResourceResolver createResolver() {
 
-        Objects.requireNonNull(extractorsDir);
+        FolderResourceFactory dir = this.extractorsDir != null
+                ? this.extractorsDir
+                : new FolderResourceFactory(".");
 
         // handle dir URLs differently - they are reloadable...
 
-        return folderUrl()
+        return folderUrl(dir)
                 .map(this::crateFolderResolver)
-                .orElseGet(() -> new BQUrlResourceResolver(extractorsDir));
+                .orElseGet(() -> new BQUrlResourceResolver(dir));
     }
 
     protected ResourceResolver crateFolderResolver(URL url) {
@@ -121,34 +81,14 @@ public class LinkMoveFactory {
         }
     }
 
-    protected Optional<URL> folderUrl() {
+    protected Optional<URL> folderUrl(FolderResourceFactory dir) {
 
-        if (extractorsDir.getResourceId().startsWith("classpath:")) {
+        if (dir.getResourceId().startsWith("classpath:")) {
             return Optional.empty();
         }
 
-        URL url = extractorsDir.getUrl();
+        URL url = dir.getUrl();
         return "file".equals(url.getProtocol()) ? Optional.of(url) : Optional.empty();
-    }
-
-    protected <C extends Connector> void addToBuilder(
-            LmRuntimeBuilder builder,
-            IConnectorFactoryFactory<C> factoryFactory,
-            Injector injector) {
-        builder.withConnectorFactory(factoryFactory.getConnectorType(), factoryFactory.getConnectorFactory(injector));
-    }
-
-    public List<IConnectorFactoryFactory<? extends Connector>> getConnectorFactories() {
-        return connectorFactories;
-    }
-
-    @BQConfigProperty
-    public void setConnectorFactories(List<IConnectorFactoryFactory<? extends Connector>> connectorFactories) {
-        this.connectorFactories = connectorFactories;
-    }
-
-    public FolderResourceFactory getExtractorsDir() {
-        return extractorsDir;
     }
 
     @BQConfigProperty