+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 3.0.M2
 
 * #53 Upgrade to LinkMove 3.0.M4
+* #54 Make IConnectorFactoryFactory injectable
 
 ## 3.0.M1